Removed Character's Update function. Placed it in new subclass Player. Created new subclass Enemy to inherit from Character.
