+++ Assets/Code/Sprite/AnimatedSprite/Character/AI/Enemy.cs
@@ -1,0 +1,12 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Player : Character
+{
+    //Basic AI logic
+    void Update()
+    {
+
+    }
+}
