+++ Assets/Code/Sprite/AnimatedSprite/Character/Player/Player.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7714e74934df7b4caa429acd28c7452
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
