+++ Assets/Prefabs/MM_Idle_01.prefab
@@ -9,8 +9,8 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 9159681931504109252}
-  - component: {fileID: 2158058944384150925}
   - component: {fileID: -6274516465647120113}
+  - component: {fileID: -6232108760227615246}
   m_Layer: 8
   m_Name: MM_Idle_01
   m_TagString: Untagged
@@ -32,26 +32,6 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &2158058944384150925
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 9159681931504109254}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 79bf086526f38cd4fa298d64a9d50460, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  spriteRenderer: {fileID: -6274516465647120113}
-  idle:
-  - {fileID: 2134549576617391521, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
-  - {fileID: -6220091244538012737, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
-  - {fileID: -1872998540497071905, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
-  - {fileID: -6041534505879322548, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
-  - {fileID: -1872998540497071905, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
-  - {fileID: -6220091244538012737, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
 --- !u!212 &-6274516465647120113
 SpriteRenderer:
   m_ObjectHideFlags: 0
@@ -89,7 +69,7 @@ SpriteRenderer:
   m_AutoUVMaxAngle: 89
   m_LightmapParameters: {fileID: 0}
   m_SortingLayerID: -652293303
-  m_SortingLayer: 1
+  m_SortingLayer: 2
   m_SortingOrder: 0
   m_Sprite: {fileID: 2134549576617391521, guid: b986ca459ab323d43a1c567762f214c9,
     type: 3}
@@ -103,3 +83,23 @@ SpriteRenderer:
   m_WasSpriteAssigned: 1
   m_MaskInteraction: 0
   m_SpriteSortPoint: 0
+--- !u!114 &-6232108760227615246
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9159681931504109254}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: c7714e74934df7b4caa429acd28c7452, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  spriteRenderer: {fileID: -6274516465647120113}
+  idle:
+  - {fileID: 2134549576617391521, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
+  - {fileID: -6220091244538012737, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
+  - {fileID: -1872998540497071905, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
+  - {fileID: -6041534505879322548, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
+  - {fileID: -1872998540497071905, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
+  - {fileID: -6220091244538012737, guid: b986ca459ab323d43a1c567762f214c9, type: 3}
